Updating validate to manage through the uploaded object's username an…

+++ server/routes/establishments/bulkUpload.js
@@ -88,15 +88,17 @@ router.route('/uploaded').post(async function (req, res) {
       });
     });
 
-    // now delete the objects in one go
-    const deleteParams = {
-      Bucket: appConfig.get('bulkuploaduser.bucketname').toString(), 
-      Delete: {
-        Objects: deleteKeys,
-        Quiet: true,
-      },
-    };
-    await s3.deleteObjects(deleteParams).promise();
+    if (deleteKeys.length > 0) {
+      // now delete the objects in one go
+      const deleteParams = {
+        Bucket: appConfig.get('bulkuploaduser.bucketname').toString(), 
+        Delete: {
+          Objects: deleteKeys,
+          Quiet: true,
+        },
+      };
+      await s3.deleteObjects(deleteParams).promise();
+    }
 
     uploadedFiles.forEach(thisFile => {
       if (thisFile.filename) {
@@ -205,18 +207,18 @@ router.route('/validate').put(async (req, res) => {
             myDownloads.establishments = myfile.data;
             establishmentMetadata.filename = myfile.filename;
             establishmentMetadata.fileType = 'Establishment';
-            
-                          
+            establishmentMetadata.userName = myfile.username;
           } else if (workerRegex.test(myfile.data.substring(0,50))) {
             myDownloads.workers = myfile.data;
             workerMetadata.filename = myfile.filename;
-            workerMetadata.fileType = 'Worker';    
-
+            workerMetadata.fileType = 'Worker';
+            workerMetadata.userName = myfile.username;
           } else if (trainingRegex.test(myfile.data.substring(0,50))) {
             myDownloads.trainings = myfile.data;
             trainingMetadata.filename = myfile.filename;
             trainingMetadata.fileType = 'Training';
-          }        
+            trainingMetadata.userName = myfile.username;
+          }
         })
     }).catch(err => {
         console.error("NM: validate.put", err);
@@ -363,10 +365,11 @@ async function downloadContent(key) {
     const filenameRegex=/^(.+\/)*(.+)\.(.+)$/; 
     
     try {
-      const objData = await s3.getObject(params).promise();      
+      const objData = await s3.getObject(params).promise();
       return {
         data: objData.Body.toString(), 
         filename: key.match(filenameRegex)[2]+ '.' + key.match(filenameRegex)[3],
+        username: objData.Metadata.username,
      };
 
     } catch (err) {
